# apps/web/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fator R API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/inbox/pgdas": {
      "post": {
        "tags": ["inbox PGDAS-D"],
        "summary": "Enviar Extrato",
        "operationId": "enviar_extrato_inbox_pgdas_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enviar_extrato_inbox_pgdas_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inbox": {
      "get": {
        "tags": ["inbox PGDAS-D"],
        "summary": "Listar Documentos",
        "operationId": "listar_documentos_inbox_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": ["received", "parsed", "needs_review", "linked", "rejected"],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentoList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inbox/{document_id}": {
      "get": {
        "tags": ["inbox PGDAS-D"],
        "summary": "Obter Documento",
        "operationId": "obter_documento_inbox__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inbox/{document_id}/arquivo": {
      "get": {
        "tags": ["inbox PGDAS-D"],
        "summary": "Baixar Original",
        "operationId": "baixar_original_inbox__document_id__arquivo_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inbox/{document_id}/link": {
      "post": {
        "tags": ["inbox PGDAS-D"],
        "summary": "Vincular Documento",
        "operationId": "vincular_documento_inbox__document_id__link_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VinculoIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inbox/{document_id}/reject": {
      "post": {
        "tags": ["inbox PGDAS-D"],
        "summary": "Rejeitar Documento",
        "operationId": "rejeitar_documento_inbox__document_id__reject_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejeicaoIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inbox/{document_id}/cadastrar-empresa": {
      "post": {
        "tags": ["inbox PGDAS-D"],
        "summary": "Cadastrar Empresa Pelo Extrato",
        "description": "Cadastra a empresa do extrato e vincula o documento, com confirmação do analista (M8).",
        "operationId": "cadastrar_empresa_pelo_extrato_inbox__document_id__cadastrar_empresa_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CadastroPeloExtratoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_enviar_extrato_inbox_pgdas_post": {
        "properties": {
          "arquivo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Arquivo"
          }
        },
        "type": "object",
        "required": ["arquivo"],
        "title": "Body_enviar_extrato_inbox_pgdas_post"
      },
      "CadastroPeloExtratoOut": {
        "properties": {
          "documento": {
            "$ref": "#/components/schemas/DocumentoOut"
          },
          "empresa": {
            "$ref": "#/components/schemas/CompanyOut"
          }
        },
        "type": "object",
        "required": ["documento", "empresa"],
        "title": "CadastroPeloExtratoOut"
      },
...
      "CompanyIn": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nome"
          },
          "cnpj": {
            "type": "string",
            "title": "Cnpj"
          },
          "cnae": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnae"
          },
          "atividade": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Atividade"
          },
          "sujeita_fator_r": {
            "type": "boolean",
            "title": "Sujeita Fator R"
          },
          "qtd_socios": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Qtd Socios",
            "default": 1
          },
          "contato": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Contato"
          },
          "pacote": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["monitoramento", "correcao", "retainer"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Pacote"
          },
          "honorario_mensal": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Honorario Mensal",
            "default": "0"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo",
            "default": true
          },
          "notas": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notas"
          },
          "inicio_atividade": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}-\\d{2}$",
                "examples": ["2026-09"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Inicio Atividade"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["nome", "cnpj", "sujeita_fator_r"],
        "title": "CompanyIn"
      },
...
      "CompanyOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "cnpj": {
            "type": "string",
            "title": "Cnpj"
          },
          "cnpj_formatado": {
            "type": "string",
            "title": "Cnpj Formatado"
          },
          "cnae": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnae"
          },
          "atividade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atividade"
          },
          "sujeita_fator_r": {
            "type": "boolean",
            "title": "Sujeita Fator R"
          },
          "qtd_socios": {
            "type": "integer",
            "title": "Qtd Socios"
          },
          "contato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contato"
          },
          "pacote": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["monitoramento", "correcao", "retainer"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Pacote"
          },
          "honorario_mensal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Honorario Mensal"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          },
          "notas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notas"
          },
          "inicio_atividade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inicio Atividade"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "cnpj",
          "cnpj_formatado",
          "cnae",
          "atividade",
          "sujeita_fator_r",
          "qtd_socios",
          "contato",
          "pacote",
          "honorario_mensal",
          "ativo",
          "notas",
          "inicio_atividade"
        ],
        "title": "CompanyOut"
      },
...
      "DocumentoList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentoOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "contagem_por_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Contagem Por Status"
          }
        },
        "type": "object",
        "required": ["items", "total", "contagem_por_status"],
        "title": "DocumentoList"
      },
      "DocumentoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "texto_agente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Texto Agente"
          },
          "status": {
            "type": "string",
            "enum": ["received", "parsed", "needs_review", "linked", "rejected"],
            "title": "Status"
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "nome_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome Original"
          },
          "mime": {
            "type": "string",
            "enum": ["application/pdf", "text/plain"],
            "title": "Mime"
          },
          "tamanho_bytes": {
            "type": "integer",
            "title": "Tamanho Bytes"
          },
          "campos": {
            "additionalProperties": true,
            "type": "object",
            "title": "Campos"
          },
          "confianca": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confianca"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "sugestao_cadastro": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SugestaoCadastroOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "motivo",
          "company_id",
          "nome_original",
          "mime",
          "tamanho_bytes",
          "campos",
          "confianca",
          "parser_version",
          "trace_id",
          "criado_em"
        ],
        "title": "DocumentoOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RejeicaoIn": {
        "properties": {
          "motivo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Motivo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["motivo"],
        "title": "RejeicaoIn"
      },
...
      "SugestaoCadastroOut": {
        "properties": {
          "cnpj": {
            "type": "string",
            "title": "Cnpj"
          },
          "cnpj_formatado": {
            "type": "string",
            "title": "Cnpj Formatado"
          },
          "nome_empresarial": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome Empresarial"
          },
          "sujeita_fator_r": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sujeita Fator R"
          },
          "inicio_atividade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inicio Atividade"
          }
        },
        "type": "object",
        "required": ["cnpj", "cnpj_formatado", "nome_empresarial", "sujeita_fator_r"],
        "title": "SugestaoCadastroOut",
        "description": "Dados do extrato para pré-preencher o cadastro de uma empresa fora da carteira (M8)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "VinculoIn": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["company_id"],
        "title": "VinculoIn"
      }
    }
  }
}